refactor: remove barrel exports from backend modules
Replace all index.ts barrel exports with direct imports to specific
files across 7 backend modules. Barrel exports provided no benefit in
a Node.js backend and led to inconsistent import patterns.

Closes #58
