--- apps/backend/src/modules/auth/index.ts
@@ -1,2 +1,0 @@
-export * from "./auth.routes.js";
-export * from "./auth.service.js";
