+++ apps/backend/src/modules/categories/category.service.test.ts
@@ -13,7 +13,7 @@ import { ConflictError, NotFoundError } from "@/common/errors.js";
 import { categoryCache } from "@/modules/categories/category.cache.js";
 import type { CategoryModelType } from "@/modules/categories/category.model.js";
 import { createCategoryService } from "@/modules/categories/category.service.js";
-import type { RecipeModelType } from "@/modules/recipes/index.js";
+import type { RecipeModelType } from "@/modules/recipes/recipe.model.js";
 
 describe("categoryService", () => {
   const categoryModel = createMockCategoryModel();
