--- apps/backend/src/modules/categories/index.ts
@@ -1,5 +1,0 @@
-export * from "./category.cache.js";
-export * from "./category.model.js";
-export * from "./category.routes.js";
-export * from "./category.schema.js";
-export * from "./category.service.js";
