--- apps/backend/src/modules/comments/index.ts
@@ -1,6 +1,0 @@
-// Routes are registered within recipe.routes.ts, not as a standalone module.
-// This is intentional: comments are a child resource of recipes.
-
-export * from "./comment.model.js";
-export * from "./comment.schema.js";
-export * from "./comment.service.js";
