--- apps/backend/src/modules/favorites/index.ts
@@ -1,4 +1,0 @@
-export * from "./favorite.model.js";
-export * from "./favorite.routes.js";
-export * from "./favorite.schema.js";
-export * from "./favorite.service.js";
