--- apps/backend/src/modules/recipe-ratings/index.ts
@@ -1,3 +1,0 @@
-export * from "./recipe-rating.model.js";
-export * from "./recipe-rating.routes.js";
-export * from "./recipe-rating.service.js";
