--- apps/backend/src/modules/recipes/index.ts
@@ -1,6 +1,0 @@
-export * from "./recipe.aggregation.js";
-export * from "./recipe.cache.js";
-export * from "./recipe.model.js";
-export * from "./recipe.routes.js";
-export * from "./recipe.schema.js";
-export * from "./recipe.service.js";
