--- apps/backend/src/modules/users/index.ts
@@ -1,3 +1,0 @@
-export * from "./user.model.js";
-export * from "./user.routes.js";
-export * from "./user.service.js";
